refactor: fix .css imports in components
