+++ frontend/src/components/ConfirmationDialog/ConfirmationDialog.tsx
@@ -1,4 +1,4 @@
-import "./style.css";
+import "./ConfirmationDialog.css";
 
 interface ConfirmDialogProps {
   open: boolean;
