+++ frontend/src/components/RoomModal/RoomModal.tsx
@@ -2,7 +2,7 @@ import { X } from "lucide-react";
 import { useCallback, useRef, useState } from "react";
 import ConfirmationDialog from "../ConfirmationDialog/ConfirmationDialog";
 import RoomForm from "./RoomForm/RoomForm";
-import "./style.css";
+import "./RoomModal.css";
 
 interface RoomModalProps {
   onClose: () => void;
