+++ frontend/src/components/Sidepanel/SidePanel.tsx
@@ -1,8 +1,8 @@
 import { motion } from "motion/react";
 import { EXPAND_COLLAPSE_TRANSITION } from "../../utils/motionTransitions";
-import RoomInfo from "./RoomInfo";
-import RoomPeople from "./RoomPeople";
-import "./style.css";
+import RoomInfo from "./RoomInfo/RoomInfo";
+import RoomPeople from "./RoomPeople/RoomPeople";
+import "./SidePanel.css";
 
 type SidePanelProps = {
   onRoomUpdate: () => Promise<void>;
