+++ frontend/src/components/TopBar/TopBar.tsx
@@ -2,8 +2,8 @@ import { Settings, User } from "lucide-react";
 import { useEffect, useState } from "react";
 import PersonSearch from "../PersonSearch/PersonSearch";
 import SettingsModal from "./SettingsModal/SettingsModal";
-import "./style.css";
-import TopBarMenu from "./TopBarMenu";
+import "./TopBar.css";
+import TopBarMenu from "./TopBarMenu/TopBarMenu";
 
 const ROOM_LABEL_FONT_SIZE = 24;
